feat: add fresh pkg to hot load; feat: add get article api; refactor: change int8,unit32 to int for valid id = -1;

+++ internal/dao/tag.go
@@ -5,18 +5,18 @@ import (
 	"github.com/iamzhiyudong/xigua-blog/pkg/app"
 )
 
-func (d *Dao) CountTag(name string, state int8) (int, error) {
+func (d *Dao) CountTag(name string, state int) (int, error) {
 	tag := model.Tag{Name: name, State: state}
 	return tag.Count(d.engine)
 }
 
-func (d *Dao) GetTagList(name string, state int8, page, pageSize int) ([]*model.Tag, error) {
+func (d *Dao) GetTagList(name string, state int, page, pageSize int) ([]*model.Tag, error) {
 	tag := model.Tag{Name: name, State: state}
 	pageOffset := app.GetPageOffset(page, pageSize)
 	return tag.List(d.engine, pageOffset, pageSize)
 }
 
-func (d *Dao) CreateTag(name string, state int8, createdBy string) error {
+func (d *Dao) CreateTag(name string, state int, createdBy string) error {
 	tag := model.Tag{
 		Name:  name,
 		State: state,
@@ -26,7 +26,7 @@ func (d *Dao) CreateTag(name string, state int8, createdBy string) error {
 	return tag.Create(d.engine)
 }
 
-func (d *Dao) UpdateTag(id uint32, name string, state int8, modifiedBy string) error {
+func (d *Dao) UpdateTag(id int, name string, state int, modifiedBy string) error {
 	tag := model.Tag{
 		Model: &model.Model{ID: id},
 	}
@@ -43,7 +43,7 @@ func (d *Dao) UpdateTag(id uint32, name string, state int8, modifiedBy string) e
 	return tag.Update(d.engine, values)
 }
 
-func (d *Dao) DeleteTag(id uint32) error {
+func (d *Dao) DeleteTag(id int) error {
 	tag := model.Tag{Model: &model.Model{ID: id}}
 	return tag.Delete(d.engine)
 }
