+++ internal/model/tag.go
@@ -8,7 +8,7 @@ import (
 type Tag struct {
 	*Model
 	Name  string `json:"name"`
-	State int8   `json:"state"`
+	State int    `json:"state"`
 }
 
 type TagSwagger struct {
